requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-wtf==1.2.1
flask-login==0.6.3
email-validator==2.1.1
pandas==2.2.1
python-dotenv==1.0.1
wtforms==3.1.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
pytest==8.0.2
mypy==1.8.0
flake8==7.0.0
black==24.2.0
azure-storage-blob==12.24.1
pyjwt==2.8.0
pytz==2024.1
psutil==7.0.0